feat(be/cover_module): add new cover field to JigData and CourseData
