+++ backend/api/src/db/jig.rs
@@ -6,12 +6,12 @@ use shared::domain::{
     asset::{DraftOrLive, PrivacyLevel},
     category::CategoryId,
     jig::{
-        module::{body::ThemeId, ModuleId},
         AudioBackground, AudioEffects, AudioFeedbackNegative, AudioFeedbackPositive,
-        DeleteUserJigs, JigAdminData, JigData, JigFocus, JigId, JigPlayerSettings, JigRating,
-        JigResponse, LiteModule, ModuleKind, TextDirection,
+        DeleteUserJigs, DraftOrLive, JigAdminData, JigData, JigFocus, JigId, JigPlayerSettings,
+        JigRating, JigResponse, PrivacyLevel, TextDirection,
     },
     meta::{AffiliationId, AgeRangeId, ResourceTypeId as TypeId},
+    module::{body::ThemeId, LiteModule, ModuleId, ModuleKind},
     user::UserScope,
 };
 use sqlx::{types::Json, PgConnection, PgPool};
@@ -221,12 +221,18 @@ select cte.jig_id                                          as "jig_id: JigId",
        curated,
        array(select row (unnest(audio_feedback_positive))) as "audio_feedback_positive!: Vec<(AudioFeedbackPositive,)>",
        array(select row (unnest(audio_feedback_negative))) as "audio_feedback_negative!: Vec<(AudioFeedbackNegative,)>",
+       (
+                select row(jig_data_module.id, kind, is_complete) 
+                from jig_data_module                
+                where jig_data_id = cte.draft_or_live_id and "index" = 0 
+                order by "index"
+       )                                                   as "cover?: (ModuleId, ModuleKind, bool)",
        array(
                select row (jig_data_module.id, kind, is_complete)
                from jig_data_module
-               where jig_data_id = cte.draft_or_live_id
+               where jig_data_id = cte.draft_or_live_id and "index" <> 0
                order by "index"
-           )                                               as "modules!: Vec<(ModuleId, ModuleKind, bool)>",
+       )                                               as "modules!: Vec<(ModuleId, ModuleKind, bool)>",
        array(select row (category_id)
              from jig_data_category
              where jig_data_id = cte.draft_or_live_id)     as "categories!: Vec<(CategoryId,)>",
@@ -262,6 +268,11 @@ from jig_data
             draft_or_live,
             display_name: row.display_name,
             language: row.language,
+            cover: row.cover.map(|(id, kind, is_complete)| LiteModule {
+                id,
+                kind,
+                is_complete,
+            }),
             modules: row
                 .modules
                 .into_iter()
@@ -390,10 +401,16 @@ select id,
        audio_background                                                              as "audio_background!: Option<AudioBackground>",
        array(select row (unnest(audio_feedback_positive)))                           as "audio_feedback_positive!: Vec<(AudioFeedbackPositive,)>",
        array(select row (unnest(audio_feedback_negative)))                           as "audio_feedback_negative!: Vec<(AudioFeedbackNegative,)>",
+       (
+                select row(jig_data_module.id, kind, is_complete) 
+                from jig_data_module                
+                where jig_data_id = jig_data.id and "index" = 0
+                order by "index"
+        )                                                   as "cover?: (ModuleId, ModuleKind, bool)",
        array(
                select row (jig_data_module.id, kind, is_complete)
                from jig_data_module
-               where jig_data_id = jig_data.id
+               where jig_data_id = jig_data.id and "index" <> 0
                order by "index"
            )                                               as "modules!: Vec<(ModuleId, ModuleKind, bool)>",
        array(select row (category_id)
@@ -440,6 +457,13 @@ from jig_data
                 draft_or_live,
                 display_name: jig_data_row.display_name,
                 language: jig_data_row.language,
+                cover: jig_data_row
+                    .cover
+                    .map(|(id, kind, is_complete)| LiteModule {
+                        id,
+                        kind,
+                        is_complete,
+                    }),
                 modules: jig_data_row
                     .modules
                     .into_iter()
@@ -587,12 +611,18 @@ select jig.id                                              as "jig_id: JigId",
    draft_or_live                                                                 as "draft_or_live!: DraftOrLive",
    array(select row (unnest(audio_feedback_positive)))                           as "audio_feedback_positive!: Vec<(AudioFeedbackPositive,)>",
    array(select row (unnest(audio_feedback_negative)))                           as "audio_feedback_negative!: Vec<(AudioFeedbackNegative,)>",
+   (
+            select row(jig_data_module.id, kind, is_complete) 
+            from jig_data_module                
+            where jig_data_id = jig_data.id and "index" = 0
+            order by "index"
+   )                                                   as "cover?: (ModuleId, ModuleKind, bool)",
    array(
            select row (jig_data_module.id, kind, is_complete)
            from jig_data_module
-           where jig_data_id = jig_data.id
+           where jig_data_id = jig_data.id and "index" <> 0
            order by "index"
-       )                                               as "modules!: Vec<(ModuleId, ModuleKind, bool)>",
+    )                                               as "modules!: Vec<(ModuleId, ModuleKind, bool)>",
    array(select row (category_id)
          from jig_data_category
          where jig_data_id = jig_data.id)     as "categories!: Vec<(CategoryId,)>",
@@ -656,6 +686,13 @@ limit $7
                 draft_or_live: jig_data_row.draft_or_live,
                 display_name: jig_data_row.display_name,
                 language: jig_data_row.language,
+                cover: jig_data_row
+                    .cover
+                    .map(|(id, kind, is_complete)| LiteModule {
+                        id,
+                        kind,
+                        is_complete,
+                    }),
                 modules: jig_data_row
                     .modules
                     .into_iter()
@@ -1151,8 +1188,8 @@ returning id
     sqlx::query!(
         //language=SQL
         r#"
-insert into jig_data_module (stable_id, "index", jig_data_id, kind, is_complete, contents)
-select stable_id, "index", $2 as "jig_id", kind, is_complete, contents
+insert into jig_data_module ("index", jig_data_id, kind, is_complete, contents)
+select "index", $2 as "jig_id", kind, is_complete, contents
 from jig_data_module
 where jig_data_id = $1
         "#,
