+++ backend/api/src/http.rs
@@ -253,8 +253,8 @@ pub fn build(
             .configure(endpoints::audio::configure)
             .configure(endpoints::meta::configure)
             .configure(endpoints::jig::configure)
-            .configure(endpoints::jig::module::configure)
-            .configure(endpoints::additional_resource::configure)
+            .configure(endpoints::module::configure)
+            .configure(endpoints::jig::additional_resource::configure)
             .configure(endpoints::jig::report::configure)
             .configure(endpoints::jig::curation::configure)
             .configure(endpoints::course::configure)
