+++ backend/api/src/http/endpoints/mod.rs
@@ -10,6 +10,7 @@ pub mod jig;
 pub mod locale;
 pub mod media;
 pub mod meta;
+pub mod module;
 pub mod pdf;
 pub mod scheduler;
 pub mod search;
