+++ backend/api/tests/integration/snapshots/integration__jig__create_default-2.snap.new
@@ -1,0 +1,50 @@
+---
+source: tests/integration/jig.rs
+expression: body
+
+---
+{
+  "id": "[id]",
+  "publishedAt": null,
+  "creatorId": "1f241e1b-b537-493f-a230-075cb16315be",
+  "authorId": "1f241e1b-b537-493f-a230-075cb16315be",
+  "authorName": "Bobby Tables",
+  "likes": 0,
+  "plays": 0,
+  "jigFocus": "modules",
+  "jigData": {
+    "draftOrLive": "draft",
+    "displayName": "",
+    "cover": null,
+    "modules": [],
+    "ageRanges": [],
+    "affiliations": [],
+    "language": "en_US",
+    "categories": [],
+    "additionalResources": [],
+    "description": "",
+    "lastEdited": "[last_edited]",
+    "defaultPlayerSettings": {
+      "direction": "ltr",
+      "displayScore": false,
+      "trackAssessments": false,
+      "dragAssist": false
+    },
+    "theme": "Blank",
+    "audioBackground": null,
+    "audioEffects": {
+      "feedbackPositive": "[audio]",
+      "feedbackNegative": "[audio]"
+    },
+    "privacyLevel": "unlisted",
+    "locked": false,
+    "otherKeywords": "",
+    "translatedKeywords": "",
+    "translatedDescription": {}
+  },
+  "adminData": {
+    "rating": null,
+    "blocked": false,
+    "curated": false
+  }
+}
