+++ shared/rust/src/domain.rs
@@ -44,6 +44,7 @@ pub mod jig;
 pub mod locale;
 pub mod media;
 pub mod meta;
+pub mod module;
 pub mod pdf;
 pub mod search;
 mod ser;
