+++ shared/rust/src/domain/module/body.rs
