+++ shared/rust/src/domain/module/body/legacy/design.rs
