+++ shared/rust/src/domain/module/body/legacy/slide.rs
@@ -2,7 +2,7 @@ pub use super::*;
 use serde::{Deserialize, Serialize};
 use serde_with::skip_serializing_none;
 
-/// The body for [`Legacy`](crate::domain::jig::module::ModuleKind::Legacy) modules.
+/// The body for [`Legacy`](crate::domain::module::ModuleKind::Legacy) modules.
 #[skip_serializing_none]
 #[derive(Default, Clone, Serialize, Deserialize, Debug)]
 pub struct Slide {
