+++ shared/rust/src/domain/module/body/tapping_board/play_settings.rs
